Fixed exec and overlay
Exec commands now more reliable

Also found that appname from googleremote is sent
when an overlay like keyboard is in-use. However,
that data isn't decodable yet, but it caused the
driver to crash. Now it handles this and does a
hex dump for future development.

This could be used to trigger keyboard interface
on the remote.